Bump units-of-measure to v5.2.1

+++ architecture/jvmfrc/src/main/kotlin/com/lynbrookrobotics/kapuchin/preferences/actual.kt
@@ -13,7 +13,7 @@ actual fun Named.pref(fallback: Float) = Preference(this, fallback, impl::putFlo
 actual fun Named.pref(fallback: Int) = Preference(this, fallback, impl::putInt, impl::getInt)
 actual fun Named.pref(fallback: Long) = Preference(this, fallback, impl::putLong, impl::getLong)
 actual fun <Q : Quan<Q>> Named.pref(fallback: Number, withUnits: UomConverter<Q>) = Preference(
-        this, withUnits(fallback),
+        this, withUnits(fallback.toDouble()),
         { name, value -> impl.putDouble(name, withUnits(value)) },
         { name, value -> withUnits(impl.getDouble(name, withUnits(value))) },
         " (${withUnits.unitName})"
